Password changed (#3)
* First Page

I implemented the First Page of our Application, which offers the possibility to log in, sing up or reset password to the user.

* Creating the files for the password reset

* Reset Password Page Done

The design of the page is done with the validation for new password and email.

* Password Changed Succesfully Page

I implemented the page which will be shown after the user submitted his request for changing the password.

---------

Co-authored-by: Antoniu12 <126654270+Antoniu12@users.noreply.github.com>