Admin account page (#14)
* Admin Page and orgnisation

* Functionality for Log In

Now you can log in as an admin too via First Page.

* Table with all users

I implemented the page for admin where he can see all the registered users. I also added the functionality to search for a specific username.

* Functionalities to delete, update or validate an userAccount

I implemented the functionalities to update, delete and validate for the admin page on user account objects registered on our app.

* Admin Table + Functionalities

I added a page where an admin can see the other admins of the site. I also implemented the functionalities for update, delete and add for userAccount objects.