Security + Log out for admin (#16)
* User Details on his pages

I implemented the functionality to show current data of the user who logged in.

* Security for pages + logout for admin

Security so that only logged in users can access the pages "user-home" , "edit-profile" and only logged in admins can access pages "user-table" and "admin-table". Also, I added a button which can log out the admin who is logged in.

---------

Co-authored-by: Antoniu12 <126654270+Antoniu12@users.noreply.github.com>