Feed Page + Logo redirect
I implemented the Feed Page where the logged in user can see it's friends' photos. Also, I added a message if an user has no friends and we try to give him the option to add new friends. Further more, I modified the redirect logo to go to the first page of the app if the logged in user is an admin, or to go to the feed page is the logged in user is a normal person.