Delete photo (#28)
* Photo Gallery

I implemented the functionality that a user can see it's own photos on his page. If he doesn't have any photos a message id displayed instead of the photos.

* Delete Photo + View Post Functionality

I implemented the View Post Functionality: when the user presses on one of his photos he can see the likes + comments on his post, or he can also delete the post if he doesn't like it anymore.

---------

Co-authored-by: Antoniu12 <126654270+Antoniu12@users.noreply.github.com>