Feed page + Logo redirect (#31)
* Photo Gallery

I implemented the functionality that a user can see it's own photos on his page. If he doesn't have any photos a message id displayed instead of the photos.

* Delete Photo + View Post Functionality

I implemented the View Post Functionality: when the user presses on one of his photos he can see the likes + comments on his post, or he can also delete the post if he doesn't like it anymore.

* Display photos of people

I added the functionality that user's profile picture is shown in the add friends table.

* View User's profile

Now when we view another user's profile  we can see it's photos + it's profile picture.

* Feed Page + Logo redirect

I implemented the Feed Page where the logged in user can see it's friends' photos. Also, I added a message if an user has no friends and we try to give him the option to add new friends. Further more, I modified the redirect logo to go to the first page of the app if the logged in user is an admin, or to go to the feed page is the logged in user is a normal person.

---------

Co-authored-by: Antoniu12 <126654270+Antoniu12@users.noreply.github.com>